requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.28
APScheduler>=3.10.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
requests>=2.31.0
dash>=2.14.0
plotly>=5.18.0
dash-bootstrap-components>=1.5.0
